Creando pageObject para Post y ajustando algunos test aleatorios.
